Decompress string when visiting YouTube page
This fixes issue #4 (again).

NOTE: This does not decompress anything when
downloading the video, it only affects situations
where we download the source for the YouTube
page so we can get at the URI.

This commit also contains a few breaking
changes and readability improvements, see
the diff log for details.